requirements.txt
pytrends
tweepy
beautifulsoup4
requests
openai
pillow
rembg
pytest